% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-symphony.R
\name{configure_symphony_solver}
\alias{configure_symphony_solver}
\title{This function constructs a symphony solver}
\usage{
configure_symphony_solver(verbose = FALSE)
}
\arguments{
\item{verbose}{would you like to see verbose messages?}
}
\description{
This function constructs a symphony solver
}

